backend/requirements.txt
beautifulsoup4>=4.12.2
fastapi>=0.109.0
httpx>=0.26.0
numpy>=1.24.0
Pillow>=10.0.0
pydantic-settings>=2.1.0
pydantic>=2.5.0
pytest-asyncio>=0.23.0
pytest>=7.4.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
qdrant-client>=1.16.2
ruff>=0.11.0
sentence-transformers>=2.2.0
structlog>=24.1.0
torch>=2.2.2
transformers>=4.36.0
uvicorn[standard]>=0.27.0